Add a basic MediaController shim and remove synchronization code
Synchronization by setting currentTime repeatedly does not seem to work
with Firefox. Instead, we can call play() on the MediaController to
start all the videos at the same time and hope they stay more or less in
sync. This breaks support for the MediaElement GUI controls provided by the
browser but I don't care about that.